# api.yml
openapi: '3.0.3'
info:
  title: ToDo app API
  version: '1.0'
servers:
  - url: http://localhost:5041
components:
  schemas:
    category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    task:
      type: object
      properties:
        id:
          type: integer
        idCategory:
          type: integer
        title:
          type: string
        description:
          type: string
        createdBy:
          type: integer
    user:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
    login:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
          format: password
  securitySchemes:
    basicAuth:
      type: http
      scheme: Basic
paths:
  /category:
    get:
      description: Lists all available categories
      security:
        - basicAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/category" 
        "401":
          description: Unathorized action
    post:
      description: Adds new category
      security:
        - basicAuth: []
      parameters:
      - name: name
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
        "401":
          description: Unathorized action
        "403":
          description: Category with that name already exists
    put:
      description: Changes the name of category with given id
      security:
        - basicAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/category"
      responses:
        "200":
          description: OK
        "401":
          description: Unathorized action
        "403":
          description: Category with that name already exists
        "404":
          description: Category with that id does not exist
    delete:
      description: Deletes category with given id
      security:
        - basicAuth: []
      parameters:
      - name: id
        in: query
        required: true
        schema:
          type: integer
      responses:
        "401":
          description: Unathorized action
        "200":
          description: OK
        "404":
          description: Category with that id does not exist
  /task:
    get:
      description: Lists all task or limited to one category
      security:
        - basicAuth: []
      parameters:
      - name: catId
        in: query
        required: false
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/task"
        "401":
          description: Unathorized action
        "404":
          description: Category with specified id does not exist
    post:
      description: Creates new task
      security:
        - basicAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/task"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
        "401":
          description: Unathorized action
    put:
      description: Modifies task with specified id
      security:
        - basicAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/task"
      responses:
        "200":
          description: OK
        "401":
          description: Unathorized action
        "404":
          description: Task with specified id not found or invalid category id
  /task/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
    get:
      description: Gets task with specified id
      security:
        - basicAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/task"
        "401":
          description: Unathorized action
        "404":
          description: Task with that id does not exist
    delete:
      description: Deletes task with specified id
      security:
        - basicAuth: []
      responses:
        "200":
          description: OK
        "401":
          description: Unathorized action
        "404":
          description: Task with specified id does not exist
  /user:
    get:
      description: Lists all users
      security:
        - basicAuth: []
      responses:
        "200":
          description:  OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/user"
        "401":
          description: Unathorized action
    post:
      description: Creates new user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/login"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: integer
        "403":
          description: Invalid username
  /user/{username}:
    parameters:
    - name: username
      in: path
      required: true
      schema:
        type: string
    put:
      parameters:
      - name: password
        in: query
        required: true
        schema:
          type: string
          format: password
      description: Changes user password
      security:
        - basicAuth: []
      responses:
        "200":
          description: OK
        "401":
          description: Unathorized action
        "403":
          description: Unathorized action
        "404":
          description: Invalid username
    delete:
      description: Deletes user
      security:
        - basicAuth: []
      responses:
        "200":
          description: OK
        "401":
          description: Unathorized action
        "403":
          description: Unathorized action
        "404":
          description: Invalid username
  /login:
    get:
      security:
        - basicAuth: []
      description: Logs in user
      responses:
        "200":
          description: OK
        "401":  
          description: Authentication invalid